requirements.txt
beautifulsoup4==4.4.1
choices==0.1
dj-config-url==0.1.1
dj-database-url==0.4.0
Django==1.9.6
django-filter==2.0.0
django-geoposition==0.3.0
django-jet==1.0.6
django-location-field==2.0.3
django-tastypie==0.13.3
django-wysiwyg-redactor==0.4.9
gunicorn==19.4.5
html-table-parser==0.1.0
Markdown==2.6.11
olefile==0.45.1
Pillow==4.0.0
simplejson==3.13.2
six==1.10.0
whitenoise==2.0.6
